+++ src/components/button-group.js
@@ -1,0 +1,40 @@
+import React from 'react';
+import getPassThroughProperties from '../utilities/component/get-pass-through-properties';
+import pureRenderShouldComponentUpdate from '../utilities/pure-render-should-component-update';
+
+class ButtonGroup extends React.Component {
+  shouldComponentUpdate(nextProps, nextState) {
+    return pureRenderShouldComponentUpdate(this.props, nextProps, this.state, nextState);
+  }
+
+  getCssClasses() {
+    let cssClasses = ['button-group'];
+
+    if (this.props.className) {
+      cssClasses = cssClasses.concat(this.props.className.split(' '));
+    }
+
+    return cssClasses;
+  }
+
+  render() {
+    return (
+      <div
+        className={this.getCssClasses().join(' ')}
+        {...getPassThroughProperties(this.props, ButtonGroup.propTypes)}
+      >
+        {this.props.children}
+      </div>
+    );
+  }
+}
+
+ButtonGroup.propTypes = {
+  className: React.PropTypes.string
+};
+
+ButtonGroup.defaultProps = {
+  className: null
+};
+
+export default ButtonGroup;